Group routes, for separate wagames and ctf platforms
