+++ src/lib/components/challengeCard.svelte
