+++ src/routes/(wargame_platform)/+error.svelte
