+++ src/routes/(wargame_platform)/+page.svelte
