+++ src/routes/(wargame_platform)/about/+page.svelte
