+++ src/routes/(wargame_platform)/admin/+page.svelte
