+++ src/routes/(wargame_platform)/admin/create-challenge/+page.server.ts
