+++ src/routes/(wargame_platform)/admin/create-challenge/+page.svelte
