+++ src/routes/(wargame_platform)/admin/create-ctf/+page.svelte
