+++ src/routes/(wargame_platform)/challenges/+page.server.ts
