+++ src/routes/(wargame_platform)/ctfs/+page.server.ts
