+++ src/routes/(wargame_platform)/scoreboard/+page.server.ts
