+++ src/routes/(wargame_platform)/scoreboard/+page.svelte
