+++ src/routes/(wargame_platform)/super_great_artist/+page.svelte
