+++ src/routes/(wargame_platform)/what_is_this_a_secret_challenge/+page.svelte
