+++ src/routes/(wargame_platform)/writeups/+page.svelte
