--- src/routes/+layout.svelte
@@ -1,14 +1,0 @@
-<script lang="ts">
-	import '../app.css';
-	import Navbar from '../components/navbar.svelte';
-	let { children, data } = $props();
-	let { user, translations } = data;
-</script>
-
-<Navbar {user} {translations}></Navbar>
-
-<main
-	class="bg-background-light dark:bg-background-dark text-foreground-light dark:text-foreground-dark flex min-h-[var(--main-height)] w-screen flex-col items-stretch py-[var(--nav-height)]"
->
-	{@render children()}
-</main>
